backend/pyproject.toml
[project]
name = "betstamp-odds-agent"
version = "0.1.0"
description = "AI-powered odds agent for anomaly detection, market analysis, and daily briefings"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "anthropic>=0.40.0",
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.1.0",
    "sse-starlette>=2.0.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
